pyproject.toml
[tool.poetry]
name = "chutes-api"
version = "0.0.1"
description = "ParachutesAI run platform API"
authors = ["Jon Durbin <jon@jondurbin.com>", "namoray <namoray@rayonlabs.ai>"]
license = "MIT"
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.10"
backoff = "^2.2.1"
requests = "^2.32.3"
loguru = "^0.7.2"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pybase64 = "^1.4.0"
orjson = "^3.10.10"
SQLAlchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
miniopy-async = "^1.21.1"
passlib = "^1.7.4"
python-multipart = "^0.0.16"
taskiq-redis = "^1.0.2"
fickling = "^0.1.3"
setuptools = "^75.3.0"
substrate-interface = "^1.7.10"
py-bip39-bindings = "^0.1.11"
hvac = "^2.3.0"
argon2-cffi = "^23.1.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}
async-lru = "^2.0.4"
# TODO: Pin the below version
fiber = { git = "https://github.com/rayonlabs/fiber.git" }
aiodns = "^3.2.0"
chutes = "^0.0.19"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
greenlet = "^3.1.1"
ipykernel = "^6.29.5"
typer = "^0.12.5"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

